Exposing MessageSender/Receiver (#146)
